requirements.txt
Flask
gunicorn
pytube
python-slugify
openai-whisper